% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Annotation.clientVersionRanges}
\alias{Annotation.clientVersionRanges}
\title{Annotation.clientVersionRanges Object}
\usage{
Annotation.clientVersionRanges(cfiRange = NULL, contentVersion = NULL,
  gbImageRange = NULL, gbTextRange = NULL, imageCfiRange = NULL)
}
\arguments{
\item{cfiRange}{Range in CFI format for this annotation sent by client}

\item{contentVersion}{Content version the client sent in}

\item{gbImageRange}{Range in GB image format for this annotation sent by client}

\item{gbTextRange}{Range in GB text format for this annotation sent by client}

\item{imageCfiRange}{Range in image CFI format for this annotation sent by client}
}
\value{
Annotation.clientVersionRanges object
}
\description{
Annotation.clientVersionRanges Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Selection ranges sent from the client.
}
\seealso{
Other Annotation functions: \code{\link{Annotation.currentVersionRanges}},
  \code{\link{Annotation.layerSummary}},
  \code{\link{Annotation}},
  \code{\link{mylibrary.annotations.insert}},
  \code{\link{mylibrary.annotations.update}}
}

