% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Annotation.currentVersionRanges}
\alias{Annotation.currentVersionRanges}
\title{Annotation.currentVersionRanges Object}
\usage{
Annotation.currentVersionRanges(cfiRange = NULL, contentVersion = NULL,
  gbImageRange = NULL, gbTextRange = NULL, imageCfiRange = NULL)
}
\arguments{
\item{cfiRange}{Range in CFI format for this annotation for version above}

\item{contentVersion}{Content version applicable to ranges below}

\item{gbImageRange}{Range in GB image format for this annotation for version above}

\item{gbTextRange}{Range in GB text format for this annotation for version above}

\item{imageCfiRange}{Range in image CFI format for this annotation for version above}
}
\value{
Annotation.currentVersionRanges object
}
\description{
Annotation.currentVersionRanges Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Selection ranges for the most recent content version.
}
\seealso{
Other Annotation functions: \code{\link{Annotation.clientVersionRanges}},
  \code{\link{Annotation.layerSummary}},
  \code{\link{Annotation}},
  \code{\link{mylibrary.annotations.insert}},
  \code{\link{mylibrary.annotations.update}}
}

