% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Annotation.layerSummary}
\alias{Annotation.layerSummary}
\title{Annotation.layerSummary Object}
\usage{
Annotation.layerSummary(allowedCharacterCount = NULL, limitType = NULL,
  remainingCharacterCount = NULL)
}
\arguments{
\item{allowedCharacterCount}{Maximum allowed characters on this layer, especially for the 'copy' layer}

\item{limitType}{Type of limitation on this layer}

\item{remainingCharacterCount}{Remaining allowed characters on this layer, especially for the 'copy' layer}
}
\value{
Annotation.layerSummary object
}
\description{
Annotation.layerSummary Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Annotation functions: \code{\link{Annotation.clientVersionRanges}},
  \code{\link{Annotation.currentVersionRanges}},
  \code{\link{Annotation}},
  \code{\link{mylibrary.annotations.insert}},
  \code{\link{mylibrary.annotations.update}}
}
