% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Annotation}
\alias{Annotation}
\title{Books API Objects 
Searches for books and manages your Google Books library.}
\usage{
Annotation(Annotation.clientVersionRanges = NULL,
  Annotation.currentVersionRanges = NULL, Annotation.layerSummary = NULL,
  afterSelectedText = NULL, beforeSelectedText = NULL,
  clientVersionRanges = NULL, created = NULL, currentVersionRanges = NULL,
  data = NULL, deleted = NULL, highlightStyle = NULL, id = NULL,
  layerId = NULL, layerSummary = NULL, pageIds = NULL,
  selectedText = NULL, selfLink = NULL, updated = NULL, volumeId = NULL)
}
\arguments{
\item{Annotation.clientVersionRanges}{The \link{Annotation.clientVersionRanges} object or list of objects}

\item{Annotation.currentVersionRanges}{The \link{Annotation.currentVersionRanges} object or list of objects}

\item{Annotation.layerSummary}{The \link{Annotation.layerSummary} object or list of objects}

\item{afterSelectedText}{Anchor text after excerpt}

\item{beforeSelectedText}{Anchor text before excerpt}

\item{clientVersionRanges}{Selection ranges sent from the client}

\item{created}{Timestamp for the created time of this annotation}

\item{currentVersionRanges}{Selection ranges for the most recent content version}

\item{data}{User-created data for this annotation}

\item{deleted}{Indicates that this annotation is deleted}

\item{highlightStyle}{The highlight style for this annotation}

\item{id}{Id of this annotation, in the form of a GUID}

\item{layerId}{The layer this annotation is for}

\item{layerSummary}{No description}

\item{pageIds}{Pages that this annotation spans}

\item{selectedText}{Excerpt from the volume}

\item{selfLink}{URL to this resource}

\item{updated}{Timestamp for the last time this annotation was modified}

\item{volumeId}{The volume that this annotation belongs to}
}
\value{
Annotation object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 19:30:05
filename: /Users/mark/dev/R/autoGoogleAPI/googlebooksv1.auto/R/books_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Annotation Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Annotation functions: \code{\link{Annotation.clientVersionRanges}},
  \code{\link{Annotation.currentVersionRanges}},
  \code{\link{Annotation.layerSummary}},
  \code{\link{mylibrary.annotations.insert}},
  \code{\link{mylibrary.annotations.update}}
}
