% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Annotationdata}
\alias{Annotationdata}
\title{Annotationdata Object}
\usage{
Annotationdata(annotationType = NULL, data = NULL, encoded_data = NULL,
  id = NULL, layerId = NULL, selfLink = NULL, updated = NULL,
  volumeId = NULL)
}
\arguments{
\item{annotationType}{The type of annotation this data is for}

\item{data}{No description}

\item{encoded_data}{Base64 encoded data for this annotation data}

\item{id}{Unique id for this annotation data}

\item{layerId}{The Layer id for this data}

\item{selfLink}{URL for this resource}

\item{updated}{Timestamp for the last time this data was updated}

\item{volumeId}{The volume id for this data}
}
\value{
Annotationdata object
}
\description{
Annotationdata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
