% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Annotations}
\alias{Annotations}
\title{Annotations Object}
\usage{
Annotations(items = NULL, nextPageToken = NULL, totalItems = NULL)
}
\arguments{
\item{items}{A list of annotations}

\item{nextPageToken}{Token to pass in for pagination for the next page}

\item{totalItems}{Total number of annotations found}
}
\value{
Annotations object
}
\description{
Annotations Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
