% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Annotationsdata}
\alias{Annotationsdata}
\title{Annotationsdata Object}
\usage{
Annotationsdata(items = NULL, nextPageToken = NULL, totalItems = NULL)
}
\arguments{
\item{items}{A list of Annotation Data}

\item{nextPageToken}{Token to pass in for pagination for the next page}

\item{totalItems}{The total number of volume annotations found}
}
\value{
Annotationsdata object
}
\description{
Annotationsdata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
