% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{AnnotationsSummary.layers}
\alias{AnnotationsSummary.layers}
\title{AnnotationsSummary.layers Object}
\usage{
AnnotationsSummary.layers()
}
\value{
AnnotationsSummary.layers object
}
\description{
AnnotationsSummary.layers Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other AnnotationsSummary functions: \code{\link{AnnotationsSummary}}
}
