% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{AnnotationsSummary}
\alias{AnnotationsSummary}
\title{AnnotationsSummary Object}
\usage{
AnnotationsSummary(AnnotationsSummary.layers = NULL, layers = NULL)
}
\arguments{
\item{AnnotationsSummary.layers}{The \link{AnnotationsSummary.layers} object or list of objects}

\item{layers}{No description}
}
\value{
AnnotationsSummary object
}
\description{
AnnotationsSummary Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other AnnotationsSummary functions: \code{\link{AnnotationsSummary.layers}}
}

