% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_functions.R
\docType{package}
\name{books_googleAuthR}
\alias{books_googleAuthR}
\alias{books_googleAuthR-package}
\title{Books API
Searches for books and manages your Google Books library.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 22:58:25
filename: /Users/mark/dev/R/autoGoogleAPI/googlebooksv1.auto/R/books_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/books
}
}

