% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{BooksAnnotationsRange}
\alias{BooksAnnotationsRange}
\title{BooksAnnotationsRange Object}
\usage{
BooksAnnotationsRange(endOffset = NULL, endPosition = NULL,
  startOffset = NULL, startPosition = NULL)
}
\arguments{
\item{endOffset}{The offset from the ending position}

\item{endPosition}{The ending position for the range}

\item{startOffset}{The offset from the starting position}

\item{startPosition}{The starting position for the range}
}
\value{
BooksAnnotationsRange object
}
\description{
BooksAnnotationsRange Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

