% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{BooksCloudloadingResource}
\alias{BooksCloudloadingResource}
\title{BooksCloudloadingResource Object}
\usage{
BooksCloudloadingResource(author = NULL, processingState = NULL,
  title = NULL, volumeId = NULL)
}
\arguments{
\item{author}{No description}

\item{processingState}{No description}

\item{title}{No description}

\item{volumeId}{No description}
}
\value{
BooksCloudloadingResource object
}
\description{
BooksCloudloadingResource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other BooksCloudloadingResource functions: \code{\link{cloudloading.updateBook}}
}
