% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Bookshelf}
\alias{Bookshelf}
\title{Bookshelf Object}
\usage{
Bookshelf(access = NULL, created = NULL, description = NULL, id = NULL,
  selfLink = NULL, title = NULL, updated = NULL, volumeCount = NULL,
  volumesLastUpdated = NULL)
}
\arguments{
\item{access}{Whether this bookshelf is PUBLIC or PRIVATE}

\item{created}{Created time for this bookshelf (formatted UTC timestamp with millisecond resolution)}

\item{description}{Description of this bookshelf}

\item{id}{Id of this bookshelf, only unique by user}

\item{selfLink}{URL to this resource}

\item{title}{Title of this bookshelf}

\item{updated}{Last modified time of this bookshelf (formatted UTC timestamp with millisecond resolution)}

\item{volumeCount}{Number of volumes in this bookshelf}

\item{volumesLastUpdated}{Last time a volume was added or removed from this bookshelf (formatted UTC timestamp with millisecond resolution)}
}
\value{
Bookshelf object
}
\description{
Bookshelf Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
