% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Bookshelves}
\alias{Bookshelves}
\title{Bookshelves Object}
\usage{
Bookshelves(items = NULL)
}
\arguments{
\item{items}{A list of bookshelves}
}
\value{
Bookshelves object
}
\description{
Bookshelves Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

