% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{BooksVolumesRecommendedRateResponse}
\alias{BooksVolumesRecommendedRateResponse}
\title{BooksVolumesRecommendedRateResponse Object}
\usage{
BooksVolumesRecommendedRateResponse(consistency_token = NULL)
}
\arguments{
\item{consistency_token}{No description}
}
\value{
BooksVolumesRecommendedRateResponse object
}
\description{
BooksVolumesRecommendedRateResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

