% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Category.items}
\alias{Category.items}
\title{Category.items Object}
\usage{
Category.items()
}
\value{
Category.items object
}
\description{
Category.items Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Category functions: \code{\link{Category}}
}
