% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Category}
\alias{Category}
\title{Category Object}
\usage{
Category(Category.items = NULL, items = NULL)
}
\arguments{
\item{Category.items}{The \link{Category.items} object or list of objects}

\item{items}{A list of onboarding categories}
}
\value{
Category object
}
\description{
Category Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Category functions: \code{\link{Category.items}}
}
