% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_functions.R
\name{cloudloading.addBook}
\alias{cloudloading.addBook}
\title{Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}}
\usage{
cloudloading.addBook(drive_document_id = NULL, mime_type = NULL,
  name = NULL, upload_client_token = NULL)
}
\arguments{
\item{drive_document_id}{A drive document id}

\item{mime_type}{The document MIME type}

\item{name}{The document name}

\item{upload_client_token}{}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/books
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/books)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/books/docs/v1/getting_started}{Google Documentation}
}
