% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_functions.R
\name{cloudloading.deleteBook}
\alias{cloudloading.deleteBook}
\title{Remove the book and its contents}
\usage{
cloudloading.deleteBook(volumeId)
}
\arguments{
\item{volumeId}{The id of the book to be removed}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/books
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/books)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/books/docs/v1/getting_started}{Google Documentation}
}

