% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_functions.R
\name{cloudloading.updateBook}
\alias{cloudloading.updateBook}
\title{Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}}
\usage{
cloudloading.updateBook(BooksCloudloadingResource)
}
\arguments{
\item{BooksCloudloadingResource}{The \link{BooksCloudloadingResource} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/books
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/books)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/books/docs/v1/getting_started}{Google Documentation}

Other BooksCloudloadingResource functions: \code{\link{BooksCloudloadingResource}}
}
