% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{ConcurrentAccessRestriction}
\alias{ConcurrentAccessRestriction}
\title{ConcurrentAccessRestriction Object}
\usage{
ConcurrentAccessRestriction(deviceAllowed = NULL,
  maxConcurrentDevices = NULL, message = NULL, nonce = NULL,
  reasonCode = NULL, restricted = NULL, signature = NULL, source = NULL,
  timeWindowSeconds = NULL, volumeId = NULL)
}
\arguments{
\item{deviceAllowed}{Whether access is granted for this (user, device, volume)}

\item{maxConcurrentDevices}{The maximum number of concurrent access licenses for this volume}

\item{message}{Error/warning message}

\item{nonce}{Client nonce for verification}

\item{reasonCode}{Error/warning reason code}

\item{restricted}{Whether this volume has any concurrent access restrictions}

\item{signature}{Response signature}

\item{source}{Client app identifier for verification}

\item{timeWindowSeconds}{Time in seconds for license auto-expiration}

\item{volumeId}{Identifies the volume for which this entry applies}
}
\value{
ConcurrentAccessRestriction object
}
\description{
ConcurrentAccessRestriction Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
