% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_functions.R
\name{dictionary.listOfflineMetadata}
\alias{dictionary.listOfflineMetadata}
\title{Returns a list of offline dictionary metadata available}
\usage{
dictionary.listOfflineMetadata(cpksver)
}
\arguments{
\item{cpksver}{The device/version ID from which to request the data}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/books
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/books)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/books/docs/v1/getting_started}{Google Documentation}
}

