% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Dictlayerdata.common}
\alias{Dictlayerdata.common}
\title{Dictlayerdata.common Object}
\usage{
Dictlayerdata.common(title = NULL)
}
\arguments{
\item{title}{The display title and localized canonical name to use when searching for this entity on Google search}
}
\value{
Dictlayerdata.common object
}
\description{
Dictlayerdata.common Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Dictlayerdata functions: \code{\link{Dictlayerdata.dict.source}},
  \code{\link{Dictlayerdata.dict.words.derivatives.source}},
  \code{\link{Dictlayerdata.dict.words.derivatives}},
  \code{\link{Dictlayerdata.dict.words.examples.source}},
  \code{\link{Dictlayerdata.dict.words.examples}},
  \code{\link{Dictlayerdata.dict.words.senses.conjugations}},
  \code{\link{Dictlayerdata.dict.words.senses.definitions.examples.source}},
  \code{\link{Dictlayerdata.dict.words.senses.definitions.examples}},
  \code{\link{Dictlayerdata.dict.words.senses.definitions}},
  \code{\link{Dictlayerdata.dict.words.senses.source}},
  \code{\link{Dictlayerdata.dict.words.senses.synonyms.source}},
  \code{\link{Dictlayerdata.dict.words.senses.synonyms}},
  \code{\link{Dictlayerdata.dict.words.senses}},
  \code{\link{Dictlayerdata.dict.words.source}},
  \code{\link{Dictlayerdata.dict.words}},
  \code{\link{Dictlayerdata.dict}},
  \code{\link{Dictlayerdata}}
}
