% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Dictlayerdata.dict.words.senses.definitions.examples}
\alias{Dictlayerdata.dict.words.senses.definitions.examples}
\title{Dictlayerdata.dict.words.senses.definitions.examples Object}
\usage{

  Dictlayerdata.dict.words.senses.definitions.examples(Dictlayerdata.dict.words.senses.definitions.examples.source = NULL)
}
\arguments{
\item{Dictlayerdata.dict.words.senses.definitions.examples.source}{The \link{Dictlayerdata.dict.words.senses.definitions.examples.source} object or list of objects}
}
\value{
Dictlayerdata.dict.words.senses.definitions.examples object
}
\description{
Dictlayerdata.dict.words.senses.definitions.examples Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Dictlayerdata functions: \code{\link{Dictlayerdata.common}},
  \code{\link{Dictlayerdata.dict.source}},
  \code{\link{Dictlayerdata.dict.words.derivatives.source}},
  \code{\link{Dictlayerdata.dict.words.derivatives}},
  \code{\link{Dictlayerdata.dict.words.examples.source}},
  \code{\link{Dictlayerdata.dict.words.examples}},
  \code{\link{Dictlayerdata.dict.words.senses.conjugations}},
  \code{\link{Dictlayerdata.dict.words.senses.definitions.examples.source}},
  \code{\link{Dictlayerdata.dict.words.senses.definitions}},
  \code{\link{Dictlayerdata.dict.words.senses.source}},
  \code{\link{Dictlayerdata.dict.words.senses.synonyms.source}},
  \code{\link{Dictlayerdata.dict.words.senses.synonyms}},
  \code{\link{Dictlayerdata.dict.words.senses}},
  \code{\link{Dictlayerdata.dict.words.source}},
  \code{\link{Dictlayerdata.dict.words}},
  \code{\link{Dictlayerdata.dict}},
  \code{\link{Dictlayerdata}}
}

