% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Dictlayerdata.dict.words.senses}
\alias{Dictlayerdata.dict.words.senses}
\title{Dictlayerdata.dict.words.senses Object}
\usage{

  Dictlayerdata.dict.words.senses(Dictlayerdata.dict.words.senses.conjugations = NULL,
  Dictlayerdata.dict.words.senses.definitions = NULL,
  Dictlayerdata.dict.words.senses.definitions.examples = NULL,
  Dictlayerdata.dict.words.senses.definitions.examples.source = NULL,
  Dictlayerdata.dict.words.senses.source = NULL,
  Dictlayerdata.dict.words.senses.synonyms = NULL,
  Dictlayerdata.dict.words.senses.synonyms.source = NULL)
}
\arguments{
\item{Dictlayerdata.dict.words.senses.conjugations}{The \link{Dictlayerdata.dict.words.senses.conjugations} object or list of objects}

\item{Dictlayerdata.dict.words.senses.definitions}{The \link{Dictlayerdata.dict.words.senses.definitions} object or list of objects}

\item{Dictlayerdata.dict.words.senses.definitions.examples}{The \link{Dictlayerdata.dict.words.senses.definitions.examples} object or list of objects}

\item{Dictlayerdata.dict.words.senses.definitions.examples.source}{The \link{Dictlayerdata.dict.words.senses.definitions.examples.source} object or list of objects}

\item{Dictlayerdata.dict.words.senses.source}{The \link{Dictlayerdata.dict.words.senses.source} object or list of objects}

\item{Dictlayerdata.dict.words.senses.synonyms}{The \link{Dictlayerdata.dict.words.senses.synonyms} object or list of objects}

\item{Dictlayerdata.dict.words.senses.synonyms.source}{The \link{Dictlayerdata.dict.words.senses.synonyms.source} object or list of objects}
}
\value{
Dictlayerdata.dict.words.senses object
}
\description{
Dictlayerdata.dict.words.senses Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Dictlayerdata functions: \code{\link{Dictlayerdata.common}},
  \code{\link{Dictlayerdata.dict.source}},
  \code{\link{Dictlayerdata.dict.words.derivatives.source}},
  \code{\link{Dictlayerdata.dict.words.derivatives}},
  \code{\link{Dictlayerdata.dict.words.examples.source}},
  \code{\link{Dictlayerdata.dict.words.examples}},
  \code{\link{Dictlayerdata.dict.words.senses.conjugations}},
  \code{\link{Dictlayerdata.dict.words.senses.definitions.examples.source}},
  \code{\link{Dictlayerdata.dict.words.senses.definitions.examples}},
  \code{\link{Dictlayerdata.dict.words.senses.definitions}},
  \code{\link{Dictlayerdata.dict.words.senses.source}},
  \code{\link{Dictlayerdata.dict.words.senses.synonyms.source}},
  \code{\link{Dictlayerdata.dict.words.senses.synonyms}},
  \code{\link{Dictlayerdata.dict.words.source}},
  \code{\link{Dictlayerdata.dict.words}},
  \code{\link{Dictlayerdata.dict}},
  \code{\link{Dictlayerdata}}
}

