% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Discoveryclusters.clusters.banner_with_content_container}
\alias{Discoveryclusters.clusters.banner_with_content_container}
\title{Discoveryclusters.clusters.banner_with_content_container Object}
\usage{

  Discoveryclusters.clusters.banner_with_content_container(fillColorArgb = NULL,
  imageUrl = NULL, maskColorArgb = NULL, moreButtonText = NULL,
  moreButtonUrl = NULL, textColorArgb = NULL)
}
\arguments{
\item{fillColorArgb}{No description}

\item{imageUrl}{No description}

\item{maskColorArgb}{No description}

\item{moreButtonText}{No description}

\item{moreButtonUrl}{No description}

\item{textColorArgb}{No description}
}
\value{
Discoveryclusters.clusters.banner_with_content_container object
}
\description{
Discoveryclusters.clusters.banner_with_content_container Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Discoveryclusters functions: \code{\link{Discoveryclusters.clusters}},
  \code{\link{Discoveryclusters}}
}
