% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Discoveryclusters.clusters}
\alias{Discoveryclusters.clusters}
\title{Discoveryclusters.clusters Object}
\usage{

  Discoveryclusters.clusters(Discoveryclusters.clusters.banner_with_content_container = NULL)
}
\arguments{
\item{Discoveryclusters.clusters.banner_with_content_container}{The \link{Discoveryclusters.clusters.banner_with_content_container} object or list of objects}
}
\value{
Discoveryclusters.clusters object
}
\description{
Discoveryclusters.clusters Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Discoveryclusters functions: \code{\link{Discoveryclusters.clusters.banner_with_content_container}},
  \code{\link{Discoveryclusters}}
}
