% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Discoveryclusters}
\alias{Discoveryclusters}
\title{Discoveryclusters Object}
\usage{
Discoveryclusters(Discoveryclusters.clusters = NULL,
  Discoveryclusters.clusters.banner_with_content_container = NULL,
  clusters = NULL, totalClusters = NULL)
}
\arguments{
\item{Discoveryclusters.clusters}{The \link{Discoveryclusters.clusters} object or list of objects}

\item{Discoveryclusters.clusters.banner_with_content_container}{The \link{Discoveryclusters.clusters.banner_with_content_container} object or list of objects}

\item{clusters}{No description}

\item{totalClusters}{No description}
}
\value{
Discoveryclusters object
}
\description{
Discoveryclusters Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Discoveryclusters functions: \code{\link{Discoveryclusters.clusters.banner_with_content_container}},
  \code{\link{Discoveryclusters.clusters}}
}
