% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{DownloadAccesses}
\alias{DownloadAccesses}
\title{DownloadAccesses Object}
\usage{
DownloadAccesses(downloadAccessList = NULL)
}
\arguments{
\item{downloadAccessList}{A list of download access responses}
}
\value{
DownloadAccesses object
}
\description{
DownloadAccesses Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
