% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{DownloadAccessRestriction}
\alias{DownloadAccessRestriction}
\title{DownloadAccessRestriction Object}
\usage{
DownloadAccessRestriction(deviceAllowed = NULL, downloadsAcquired = NULL,
  justAcquired = NULL, maxDownloadDevices = NULL, message = NULL,
  nonce = NULL, reasonCode = NULL, restricted = NULL, signature = NULL,
  source = NULL, volumeId = NULL)
}
\arguments{
\item{deviceAllowed}{If restricted, whether access is granted for this (user, device, volume)}

\item{downloadsAcquired}{If restricted, the number of content download licenses already acquired (including the requesting client, if licensed)}

\item{justAcquired}{If deviceAllowed, whether access was just acquired with this request}

\item{maxDownloadDevices}{If restricted, the maximum number of content download licenses for this volume}

\item{message}{Error/warning message}

\item{nonce}{Client nonce for verification}

\item{reasonCode}{Error/warning reason code}

\item{restricted}{Whether this volume has any download access restrictions}

\item{signature}{Response signature}

\item{source}{Client app identifier for verification}

\item{volumeId}{Identifies the volume for which this entry applies}
}
\value{
DownloadAccessRestriction object
}
\description{
DownloadAccessRestriction Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

