% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Geolayerdata.common}
\alias{Geolayerdata.common}
\title{Geolayerdata.common Object}
\usage{
Geolayerdata.common(lang = NULL, previewImageUrl = NULL, snippet = NULL,
  snippetUrl = NULL, title = NULL)
}
\arguments{
\item{lang}{The language of the information url and description}

\item{previewImageUrl}{The URL for the preview image information}

\item{snippet}{The description for this location}

\item{snippetUrl}{The URL for information for this location}

\item{title}{The display title and localized canonical name to use when searching for this entity on Google search}
}
\value{
Geolayerdata.common object
}
\description{
Geolayerdata.common Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Geolayerdata functions: \code{\link{Geolayerdata.geo.viewport.hi}},
  \code{\link{Geolayerdata.geo.viewport.lo}},
  \code{\link{Geolayerdata.geo.viewport}},
  \code{\link{Geolayerdata.geo}},
  \code{\link{Geolayerdata}}
}
