% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Geolayerdata.geo}
\alias{Geolayerdata.geo}
\title{Geolayerdata.geo Object}
\usage{
Geolayerdata.geo(Geolayerdata.geo.viewport = NULL,
  Geolayerdata.geo.viewport.hi = NULL, Geolayerdata.geo.viewport.lo = NULL,
  boundary = NULL, cachePolicy = NULL, countryCode = NULL,
  latitude = NULL, longitude = NULL, mapType = NULL, viewport = NULL,
  zoom = NULL)
}
\arguments{
\item{Geolayerdata.geo.viewport}{The \link{Geolayerdata.geo.viewport} object or list of objects}

\item{Geolayerdata.geo.viewport.hi}{The \link{Geolayerdata.geo.viewport.hi} object or list of objects}

\item{Geolayerdata.geo.viewport.lo}{The \link{Geolayerdata.geo.viewport.lo} object or list of objects}

\item{boundary}{The boundary of the location as a set of loops containing pairs of latitude, longitude coordinates}

\item{cachePolicy}{The cache policy active for this data}

\item{countryCode}{The country code of the location}

\item{latitude}{The latitude of the location}

\item{longitude}{The longitude of the location}

\item{mapType}{The type of map that should be used for this location}

\item{viewport}{The viewport for showing this location}

\item{zoom}{The Zoom level to use for the map}
}
\value{
Geolayerdata.geo object
}
\description{
Geolayerdata.geo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Geolayerdata functions: \code{\link{Geolayerdata.common}},
  \code{\link{Geolayerdata.geo.viewport.hi}},
  \code{\link{Geolayerdata.geo.viewport.lo}},
  \code{\link{Geolayerdata.geo.viewport}},
  \code{\link{Geolayerdata}}
}
