% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Geolayerdata.geo.viewport.hi}
\alias{Geolayerdata.geo.viewport.hi}
\title{Geolayerdata.geo.viewport.hi Object}
\usage{
Geolayerdata.geo.viewport.hi(latitude = NULL, longitude = NULL)
}
\arguments{
\item{latitude}{No description}

\item{longitude}{No description}
}
\value{
Geolayerdata.geo.viewport.hi object
}
\description{
Geolayerdata.geo.viewport.hi Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Geolayerdata functions: \code{\link{Geolayerdata.common}},
  \code{\link{Geolayerdata.geo.viewport.lo}},
  \code{\link{Geolayerdata.geo.viewport}},
  \code{\link{Geolayerdata.geo}},
  \code{\link{Geolayerdata}}
}

