% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Geolayerdata.geo.viewport}
\alias{Geolayerdata.geo.viewport}
\title{Geolayerdata.geo.viewport Object}
\usage{
Geolayerdata.geo.viewport(Geolayerdata.geo.viewport.hi = NULL,
  Geolayerdata.geo.viewport.lo = NULL, hi = NULL, lo = NULL)
}
\arguments{
\item{Geolayerdata.geo.viewport.hi}{The \link{Geolayerdata.geo.viewport.hi} object or list of objects}

\item{Geolayerdata.geo.viewport.lo}{The \link{Geolayerdata.geo.viewport.lo} object or list of objects}

\item{hi}{No description}

\item{lo}{No description}
}
\value{
Geolayerdata.geo.viewport object
}
\description{
Geolayerdata.geo.viewport Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The viewport for showing this location. This is a latitude, longitude rectangle.
}
\seealso{
Other Geolayerdata functions: \code{\link{Geolayerdata.common}},
  \code{\link{Geolayerdata.geo.viewport.hi}},
  \code{\link{Geolayerdata.geo.viewport.lo}},
  \code{\link{Geolayerdata.geo}},
  \code{\link{Geolayerdata}}
}
