% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Geolayerdata}
\alias{Geolayerdata}
\title{Geolayerdata Object}
\usage{
Geolayerdata(Geolayerdata.common = NULL, Geolayerdata.geo = NULL,
  Geolayerdata.geo.viewport = NULL, Geolayerdata.geo.viewport.hi = NULL,
  Geolayerdata.geo.viewport.lo = NULL, common = NULL, geo = NULL)
}
\arguments{
\item{Geolayerdata.common}{The \link{Geolayerdata.common} object or list of objects}

\item{Geolayerdata.geo}{The \link{Geolayerdata.geo} object or list of objects}

\item{Geolayerdata.geo.viewport}{The \link{Geolayerdata.geo.viewport} object or list of objects}

\item{Geolayerdata.geo.viewport.hi}{The \link{Geolayerdata.geo.viewport.hi} object or list of objects}

\item{Geolayerdata.geo.viewport.lo}{The \link{Geolayerdata.geo.viewport.lo} object or list of objects}

\item{common}{No description}

\item{geo}{No description}
}
\value{
Geolayerdata object
}
\description{
Geolayerdata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Geolayerdata functions: \code{\link{Geolayerdata.common}},
  \code{\link{Geolayerdata.geo.viewport.hi}},
  \code{\link{Geolayerdata.geo.viewport.lo}},
  \code{\link{Geolayerdata.geo.viewport}},
  \code{\link{Geolayerdata.geo}}
}

