% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_functions.R
\name{layers.get}
\alias{layers.get}
\title{Gets the layer summary for a volume.}
\usage{
layers.get(volumeId, summaryId, contentVersion = NULL, source = NULL)
}
\arguments{
\item{volumeId}{The volume to retrieve layers for}

\item{summaryId}{The ID for the layer to get the summary for}

\item{contentVersion}{The content version for the requested volume}

\item{source}{String to identify the originator of this request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/books
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/books)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/books/docs/v1/getting_started}{Google Documentation}
}
