% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_functions.R
\name{layers.list}
\alias{layers.list}
\title{List the layer summaries for a volume.}
\usage{
layers.list(volumeId, contentVersion = NULL, maxResults = NULL,
  pageToken = NULL, source = NULL)
}
\arguments{
\item{volumeId}{The volume to retrieve layers for}

\item{contentVersion}{The content version for the requested volume}

\item{maxResults}{Maximum number of results to return}

\item{pageToken}{The value of the nextToken from the previous page}

\item{source}{String to identify the originator of this request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/books
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/books)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/books/docs/v1/getting_started}{Google Documentation}
}

