% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Layersummaries}
\alias{Layersummaries}
\title{Layersummaries Object}
\usage{
Layersummaries(items = NULL, totalItems = NULL)
}
\arguments{
\item{items}{A list of layer summary items}

\item{totalItems}{The total number of layer summaries found}
}
\value{
Layersummaries object
}
\description{
Layersummaries Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
