% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Layersummary}
\alias{Layersummary}
\title{Layersummary Object}
\usage{
Layersummary(annotationCount = NULL, annotationTypes = NULL,
  annotationsDataLink = NULL, annotationsLink = NULL,
  contentVersion = NULL, dataCount = NULL, id = NULL, layerId = NULL,
  selfLink = NULL, updated = NULL, volumeAnnotationsVersion = NULL,
  volumeId = NULL)
}
\arguments{
\item{annotationCount}{The number of annotations for this layer}

\item{annotationTypes}{The list of annotation types contained for this layer}

\item{annotationsDataLink}{Link to get data for this annotation}

\item{annotationsLink}{The link to get the annotations for this layer}

\item{contentVersion}{The content version this resource is for}

\item{dataCount}{The number of data items for this layer}

\item{id}{Unique id of this layer summary}

\item{layerId}{The layer id for this summary}

\item{selfLink}{URL to this resource}

\item{updated}{Timestamp for the last time an item in this layer was updated}

\item{volumeAnnotationsVersion}{The current version of this layer's volume annotations}

\item{volumeId}{The volume id this resource is for}
}
\value{
Layersummary object
}
\description{
Layersummary Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

