% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_functions.R
\name{myconfig.releaseDownloadAccess}
\alias{myconfig.releaseDownloadAccess}
\title{Release downloaded content access restriction.}
\usage{
myconfig.releaseDownloadAccess(volumeIds, cpksver, locale = NULL,
  source = NULL)
}
\arguments{
\item{volumeIds}{The volume(s) to release restrictions for}

\item{cpksver}{The device/version ID from which to release the restriction}

\item{locale}{ISO-639-1, ISO-3166-1 codes for message localization, i}

\item{source}{String to identify the originator of this request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/books
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/books)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/books/docs/v1/getting_started}{Google Documentation}
}

