% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_functions.R
\name{myconfig.requestAccess}
\alias{myconfig.requestAccess}
\title{Request concurrent and download access restrictions.}
\usage{
myconfig.requestAccess(source, volumeId, nonce, cpksver, licenseTypes = NULL,
  locale = NULL)
}
\arguments{
\item{source}{String to identify the originator of this request}

\item{volumeId}{The volume to request concurrent/download restrictions for}

\item{nonce}{The client nonce value}

\item{cpksver}{The device/version ID from which to request the restrictions}

\item{licenseTypes}{The type of access license to request}

\item{locale}{ISO-639-1, ISO-3166-1 codes for message localization, i}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/books
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/books)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/books/docs/v1/getting_started}{Google Documentation}
}

