% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_functions.R
\name{myconfig.syncVolumeLicenses}
\alias{myconfig.syncVolumeLicenses}
\title{Request downloaded content access for specified volumes on the My eBooks shelf.}
\usage{
myconfig.syncVolumeLicenses(source, nonce, cpksver, features = NULL,
  includeNonComicsSeries = NULL, locale = NULL, showPreorders = NULL,
  volumeIds = NULL)
}
\arguments{
\item{source}{String to identify the originator of this request}

\item{nonce}{The client nonce value}

\item{cpksver}{The device/version ID from which to release the restriction}

\item{features}{List of features supported by the client, i}

\item{includeNonComicsSeries}{Set to true to include non-comics series}

\item{locale}{ISO-639-1, ISO-3166-1 codes for message localization, i}

\item{showPreorders}{Set to true to show pre-ordered books}

\item{volumeIds}{The volume(s) to request download restrictions for}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/books
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/books)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/books/docs/v1/getting_started}{Google Documentation}
}

