% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_functions.R
\name{myconfig.updateUserSettings}
\alias{myconfig.updateUserSettings}
\title{Sets the settings for the user. If a sub-object is specified, it will overwrite the existing sub-object stored in the server. Unspecified sub-objects will retain the existing value.}
\usage{
myconfig.updateUserSettings(Usersettings)
}
\arguments{
\item{Usersettings}{The \link{Usersettings} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/books
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/books)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/books/docs/v1/getting_started}{Google Documentation}

Other Usersettings functions: \code{\link{Usersettings.notesExport}},
  \code{\link{Usersettings.notification.moreFromAuthors}},
  \code{\link{Usersettings.notification.moreFromSeries}},
  \code{\link{Usersettings.notification}},
  \code{\link{Usersettings}}
}

