% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_functions.R
\name{mylibrary.annotations.summary}
\alias{mylibrary.annotations.summary}
\title{Gets the summary of specified layers.}
\usage{
mylibrary.annotations.summary(layerIds, volumeId)
}
\arguments{
\item{layerIds}{Array of layer IDs to get the summary for}

\item{volumeId}{Volume id to get the summary for}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/books
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/books)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/books/docs/v1/getting_started}{Google Documentation}
}

