% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_functions.R
\name{mylibrary.annotations.update}
\alias{mylibrary.annotations.update}
\title{Updates an existing annotation.}
\usage{
mylibrary.annotations.update(Annotation, annotationId, source = NULL)
}
\arguments{
\item{Annotation}{The \link{Annotation} object to pass to this method}

\item{annotationId}{The ID for the annotation to update}

\item{source}{String to identify the originator of this request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/books
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/books)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/books/docs/v1/getting_started}{Google Documentation}

Other Annotation functions: \code{\link{Annotation.clientVersionRanges}},
  \code{\link{Annotation.currentVersionRanges}},
  \code{\link{Annotation.layerSummary}},
  \code{\link{Annotation}},
  \code{\link{mylibrary.annotations.insert}}
}

