% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_functions.R
\name{mylibrary.elves.addVolume}
\alias{mylibrary.elves.addVolume}
\title{Adds a volume to a bookshelf.}
\usage{
mylibrary.elves.addVolume(shelf, volumeId, reason = NULL, source = NULL)
}
\arguments{
\item{shelf}{ID of bookshelf to which to add a volume}

\item{volumeId}{ID of volume to add}

\item{reason}{The reason for which the book is added to the library}

\item{source}{String to identify the originator of this request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/books
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/books)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/books/docs/v1/getting_started}{Google Documentation}
}

