% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_functions.R
\name{mylibrary.elves.moveVolume}
\alias{mylibrary.elves.moveVolume}
\title{Moves a volume within a bookshelf.}
\usage{
mylibrary.elves.moveVolume(shelf, volumeId, volumePosition, source = NULL)
}
\arguments{
\item{shelf}{ID of bookshelf with the volume}

\item{volumeId}{ID of volume to move}

\item{volumePosition}{Position on shelf to move the item (0 puts the item before the current first item, 1 puts it between the first and the second and so on}

\item{source}{String to identify the originator of this request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/books
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/books)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/books/docs/v1/getting_started}{Google Documentation}
}

