% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_functions.R
\name{mylibrary.elves.removeVolume}
\alias{mylibrary.elves.removeVolume}
\title{Removes a volume from a bookshelf.}
\usage{
mylibrary.elves.removeVolume(shelf, volumeId, reason = NULL, source = NULL)
}
\arguments{
\item{shelf}{ID of bookshelf from which to remove a volume}

\item{volumeId}{ID of volume to remove}

\item{reason}{The reason for which the book is removed from the library}

\item{source}{String to identify the originator of this request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/books
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/books)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/books/docs/v1/getting_started}{Google Documentation}
}
