% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_functions.R
\name{mylibrary.readingpositions.setPosition}
\alias{mylibrary.readingpositions.setPosition}
\title{Sets my reading position information for a volume.}
\usage{
mylibrary.readingpositions.setPosition(volumeId, timestamp, position,
  action = NULL, contentVersion = NULL, deviceCookie = NULL,
  source = NULL)
}
\arguments{
\item{volumeId}{ID of volume for which to update the reading position}

\item{timestamp}{RFC 3339 UTC format timestamp associated with this reading position}

\item{position}{Position string for the new volume reading position}

\item{action}{Action that caused this reading position to be set}

\item{contentVersion}{Volume content version for which this reading position applies}

\item{deviceCookie}{Random persistent device cookie optional on set position}

\item{source}{String to identify the originator of this request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/books
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/books)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/books/docs/v1/getting_started}{Google Documentation}
}
