% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Notification}
\alias{Notification}
\title{Notification Object}
\usage{
Notification(body = NULL, crmExperimentIds = NULL, doc_id = NULL,
  doc_type = NULL, dont_show_notification = NULL, iconUrl = NULL,
  notificationGroup = NULL, notification_type = NULL, pcampaign_id = NULL,
  reason = NULL, show_notification_settings_action = NULL,
  targetUrl = NULL, title = NULL)
}
\arguments{
\item{body}{No description}

\item{crmExperimentIds}{The list of crm experiment ids}

\item{doc_id}{No description}

\item{doc_type}{No description}

\item{dont_show_notification}{No description}

\item{iconUrl}{No description}

\item{notificationGroup}{No description}

\item{notification_type}{No description}

\item{pcampaign_id}{No description}

\item{reason}{No description}

\item{show_notification_settings_action}{No description}

\item{targetUrl}{No description}

\item{title}{No description}
}
\value{
Notification object
}
\description{
Notification Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
