% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Offers.items.items}
\alias{Offers.items.items}
\title{Offers.items.items Object}
\usage{
Offers.items.items()
}
\value{
Offers.items.items object
}
\description{
Offers.items.items Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Offers functions: \code{\link{Offers.items}},
  \code{\link{Offers}}
}

