% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Offers.items}
\alias{Offers.items}
\title{Offers.items Object}
\usage{
Offers.items(Offers.items.items = NULL)
}
\arguments{
\item{Offers.items.items}{The \link{Offers.items.items} object or list of objects}
}
\value{
Offers.items object
}
\description{
Offers.items Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Offers functions: \code{\link{Offers.items.items}},
  \code{\link{Offers}}
}

