% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Offers}
\alias{Offers}
\title{Offers Object}
\usage{
Offers(Offers.items = NULL, Offers.items.items = NULL, items = NULL)
}
\arguments{
\item{Offers.items}{The \link{Offers.items} object or list of objects}

\item{Offers.items.items}{The \link{Offers.items.items} object or list of objects}

\item{items}{A list of offers}
}
\value{
Offers object
}
\description{
Offers Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Offers functions: \code{\link{Offers.items.items}},
  \code{\link{Offers.items}}
}
