% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_functions.R
\name{onboarding.listCategoryVolumes}
\alias{onboarding.listCategoryVolumes}
\title{List available volumes under categories for onboarding experience.}
\usage{
onboarding.listCategoryVolumes(categoryId = NULL, locale = NULL,
  maxAllowedMaturityRating = NULL, pageSize = NULL, pageToken = NULL)
}
\arguments{
\item{categoryId}{List of category ids requested}

\item{locale}{ISO-639-1 language and ISO-3166-1 country code}

\item{maxAllowedMaturityRating}{The maximum allowed maturity rating of returned volumes}

\item{pageSize}{Number of maximum results per page to be included in the response}

\item{pageToken}{The value of the nextToken from the previous page}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/books
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/books)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/books/docs/v1/getting_started}{Google Documentation}
}
