% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_functions.R
\name{personalizedstream.get}
\alias{personalizedstream.get}
\title{Returns a stream of personalized book clusters}
\usage{
personalizedstream.get(locale = NULL, maxAllowedMaturityRating = NULL,
  source = NULL)
}
\arguments{
\item{locale}{ISO-639-1 language and ISO-3166-1 country code}

\item{maxAllowedMaturityRating}{The maximum allowed maturity rating of returned recommendations}

\item{source}{String to identify the originator of this request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/books
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/books)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/books/docs/v1/getting_started}{Google Documentation}
}
