% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_functions.R
\name{promooffer.accept}
\alias{promooffer.accept}
\title{Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}}
\usage{
promooffer.accept(androidId = NULL, device = NULL, manufacturer = NULL,
  model = NULL, offerId = NULL, product = NULL, serial = NULL,
  volumeId = NULL)
}
\arguments{
\item{androidId}{device android_id}

\item{device}{device device}

\item{manufacturer}{device manufacturer}

\item{model}{device model}

\item{offerId}{}

\item{product}{device product}

\item{serial}{device serial}

\item{volumeId}{Volume id to exercise the offer}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/books
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/books)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/books/docs/v1/getting_started}{Google Documentation}
}

