% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{RequestAccess}
\alias{RequestAccess}
\title{RequestAccess Object}
\usage{
RequestAccess(concurrentAccess = NULL, downloadAccess = NULL)
}
\arguments{
\item{concurrentAccess}{A concurrent access response}

\item{downloadAccess}{A download access response}
}
\value{
RequestAccess object
}
\description{
RequestAccess Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

