% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Review.author}
\alias{Review.author}
\title{Review.author Object}
\usage{
Review.author(displayName = NULL)
}
\arguments{
\item{displayName}{Name of this person}
}
\value{
Review.author object
}
\description{
Review.author Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Author of this review.
}
\seealso{
Other Review functions: \code{\link{Review.source}},
  \code{\link{Review}}
}
