% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Review}
\alias{Review}
\title{Review Object}
\usage{
Review(Review.author = NULL, Review.source = NULL, author = NULL,
  content = NULL, date = NULL, fullTextUrl = NULL, rating = NULL,
  source = NULL, title = NULL, type = NULL, volumeId = NULL)
}
\arguments{
\item{Review.author}{The \link{Review.author} object or list of objects}

\item{Review.source}{The \link{Review.source} object or list of objects}

\item{author}{Author of this review}

\item{content}{Review text}

\item{date}{Date of this review}

\item{fullTextUrl}{URL for the full review text, for reviews gathered from the web}

\item{rating}{Star rating for this review}

\item{source}{Information regarding the source of this review, when the review is not from a Google Books user}

\item{title}{Title for this review}

\item{type}{Source type for this review}

\item{volumeId}{Volume that this review is for}
}
\value{
Review object
}
\description{
Review Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Review functions: \code{\link{Review.author}},
  \code{\link{Review.source}}
}

