% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Review.source}
\alias{Review.source}
\title{Review.source Object}
\usage{
Review.source(description = NULL, extraDescription = NULL, url = NULL)
}
\arguments{
\item{description}{Name of the source}

\item{extraDescription}{Extra text about the source of the review}

\item{url}{URL of the source of the review}
}
\value{
Review.source object
}
\description{
Review.source Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information regarding the source of this review, when the review is not from a Google Books user.
}
\seealso{
Other Review functions: \code{\link{Review.author}},
  \code{\link{Review}}
}

