% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_functions.R
\name{series.get}
\alias{series.get}
\title{Returns Series metadata for the given series ids.}
\usage{
series.get(series_id)
}
\arguments{
\item{series_id}{String that identifies the series}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/books
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/books)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/books/docs/v1/getting_started}{Google Documentation}
}
