% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Series}
\alias{Series}
\title{Series Object}
\usage{
Series(Series.series = NULL, series = NULL)
}
\arguments{
\item{Series.series}{The \link{Series.series} object or list of objects}

\item{series}{No description}
}
\value{
Series object
}
\description{
Series Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Series functions: \code{\link{Series.series}}
}

