% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Series.series}
\alias{Series.series}
\title{Series.series Object}
\usage{
Series.series()
}
\value{
Series.series object
}
\description{
Series.series Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Series functions: \code{\link{Series}}
}

