% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Seriesmembership}
\alias{Seriesmembership}
\title{Seriesmembership Object}
\usage{
Seriesmembership(member = NULL, nextPageToken = NULL)
}
\arguments{
\item{member}{No description}

\item{nextPageToken}{No description}
}
\value{
Seriesmembership object
}
\description{
Seriesmembership Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

