% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Usersettings.notesExport}
\alias{Usersettings.notesExport}
\title{Usersettings.notesExport Object}
\usage{
Usersettings.notesExport(folderName = NULL, isEnabled = NULL)
}
\arguments{
\item{folderName}{No description}

\item{isEnabled}{No description}
}
\value{
Usersettings.notesExport object
}
\description{
Usersettings.notesExport Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
User settings in sub-objects, each for different purposes.
}
\seealso{
Other Usersettings functions: \code{\link{Usersettings.notification.moreFromAuthors}},
  \code{\link{Usersettings.notification.moreFromSeries}},
  \code{\link{Usersettings.notification}},
  \code{\link{Usersettings}},
  \code{\link{myconfig.updateUserSettings}}
}

