% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Usersettings.notification.moreFromAuthors}
\alias{Usersettings.notification.moreFromAuthors}
\title{Usersettings.notification.moreFromAuthors Object}
\usage{
Usersettings.notification.moreFromAuthors(opted_state = NULL)
}
\arguments{
\item{opted_state}{No description}
}
\value{
Usersettings.notification.moreFromAuthors object
}
\description{
Usersettings.notification.moreFromAuthors Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Usersettings functions: \code{\link{Usersettings.notesExport}},
  \code{\link{Usersettings.notification.moreFromSeries}},
  \code{\link{Usersettings.notification}},
  \code{\link{Usersettings}},
  \code{\link{myconfig.updateUserSettings}}
}

