% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Usersettings.notification}
\alias{Usersettings.notification}
\title{Usersettings.notification Object}
\usage{
Usersettings.notification(Usersettings.notification.moreFromAuthors = NULL,
  Usersettings.notification.moreFromSeries = NULL, moreFromAuthors = NULL,
  moreFromSeries = NULL)
}
\arguments{
\item{Usersettings.notification.moreFromAuthors}{The \link{Usersettings.notification.moreFromAuthors} object or list of objects}

\item{Usersettings.notification.moreFromSeries}{The \link{Usersettings.notification.moreFromSeries} object or list of objects}

\item{moreFromAuthors}{No description}

\item{moreFromSeries}{No description}
}
\value{
Usersettings.notification object
}
\description{
Usersettings.notification Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Usersettings functions: \code{\link{Usersettings.notesExport}},
  \code{\link{Usersettings.notification.moreFromAuthors}},
  \code{\link{Usersettings.notification.moreFromSeries}},
  \code{\link{Usersettings}},
  \code{\link{myconfig.updateUserSettings}}
}

