% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Usersettings}
\alias{Usersettings}
\title{Usersettings Object}
\usage{
Usersettings(Usersettings.notesExport = NULL,
  Usersettings.notification = NULL,
  Usersettings.notification.moreFromAuthors = NULL,
  Usersettings.notification.moreFromSeries = NULL, notesExport = NULL,
  notification = NULL)
}
\arguments{
\item{Usersettings.notesExport}{The \link{Usersettings.notesExport} object or list of objects}

\item{Usersettings.notification}{The \link{Usersettings.notification} object or list of objects}

\item{Usersettings.notification.moreFromAuthors}{The \link{Usersettings.notification.moreFromAuthors} object or list of objects}

\item{Usersettings.notification.moreFromSeries}{The \link{Usersettings.notification.moreFromSeries} object or list of objects}

\item{notesExport}{User settings in sub-objects, each for different purposes}

\item{notification}{No description}
}
\value{
Usersettings object
}
\description{
Usersettings Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Usersettings functions: \code{\link{Usersettings.notesExport}},
  \code{\link{Usersettings.notification.moreFromAuthors}},
  \code{\link{Usersettings.notification.moreFromSeries}},
  \code{\link{Usersettings.notification}},
  \code{\link{myconfig.updateUserSettings}}
}

