% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Volume.accessInfo}
\alias{Volume.accessInfo}
\title{Volume.accessInfo Object}
\usage{
Volume.accessInfo(Volume.accessInfo.epub = NULL,
  Volume.accessInfo.pdf = NULL, accessViewStatus = NULL, country = NULL,
  downloadAccess = NULL, driveImportedContentLink = NULL,
  embeddable = NULL, epub = NULL, explicitOfflineLicenseManagement = NULL,
  pdf = NULL, publicDomain = NULL, quoteSharingAllowed = NULL,
  textToSpeechPermission = NULL, viewOrderUrl = NULL, viewability = NULL,
  webReaderLink = NULL)
}
\arguments{
\item{Volume.accessInfo.epub}{The \link{Volume.accessInfo.epub} object or list of objects}

\item{Volume.accessInfo.pdf}{The \link{Volume.accessInfo.pdf} object or list of objects}

\item{accessViewStatus}{Combines the access and viewability of this volume into a single status field for this user}

\item{country}{The two-letter ISO_3166-1 country code for which this access information is valid}

\item{downloadAccess}{Information about a volume's download license access restrictions}

\item{driveImportedContentLink}{URL to the Google Drive viewer if this volume is uploaded by the user by selecting the file from Google Drive}

\item{embeddable}{Whether this volume can be embedded in a viewport using the Embedded Viewer API}

\item{epub}{Information about epub content}

\item{explicitOfflineLicenseManagement}{Whether this volume requires that the client explicitly request offline download license rather than have it done automatically when loading the content, if the client supports it}

\item{pdf}{Information about pdf content}

\item{publicDomain}{Whether or not this book is public domain in the country listed above}

\item{quoteSharingAllowed}{Whether quote sharing is allowed for this volume}

\item{textToSpeechPermission}{Whether text-to-speech is permitted for this volume}

\item{viewOrderUrl}{For ordered but not yet processed orders, we give a URL that can be used to go to the appropriate Google Wallet page}

\item{viewability}{The read access of a volume}

\item{webReaderLink}{URL to read this volume on the Google Books site}
}
\value{
Volume.accessInfo object
}
\description{
Volume.accessInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Any information about a volume related to reading or obtaining that volume text. This information can depend on country (books may be public domain in one country but not in another, e.g.).
}
\seealso{
Other Volume functions: \code{\link{Volume.accessInfo.epub}},
  \code{\link{Volume.accessInfo.pdf}},
  \code{\link{Volume.layerInfo.layers}},
  \code{\link{Volume.layerInfo}},
  \code{\link{Volume.recommendedInfo}},
  \code{\link{Volume.saleInfo.listPrice}},
  \code{\link{Volume.saleInfo.offers.listPrice}},
  \code{\link{Volume.saleInfo.offers.rentalDuration}},
  \code{\link{Volume.saleInfo.offers.retailPrice}},
  \code{\link{Volume.saleInfo.offers}},
  \code{\link{Volume.saleInfo.retailPrice}},
  \code{\link{Volume.saleInfo}},
  \code{\link{Volume.searchInfo}},
  \code{\link{Volume.userInfo.copy}},
  \code{\link{Volume.userInfo.familySharing}},
  \code{\link{Volume.userInfo.rentalPeriod}},
  \code{\link{Volume.userInfo.userUploadedVolumeInfo}},
  \code{\link{Volume.userInfo}},
  \code{\link{Volume.volumeInfo.dimensions}},
  \code{\link{Volume.volumeInfo.imageLinks}},
  \code{\link{Volume.volumeInfo.industryIdentifiers}},
  \code{\link{Volume.volumeInfo.panelizationSummary}},
  \code{\link{Volume.volumeInfo}}, \code{\link{Volume}}
}

