% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Volume}
\alias{Volume}
\title{Volume Object}
\usage{
Volume(Volume.accessInfo = NULL, Volume.accessInfo.epub = NULL,
  Volume.accessInfo.pdf = NULL, Volume.layerInfo = NULL,
  Volume.layerInfo.layers = NULL, Volume.recommendedInfo = NULL,
  Volume.saleInfo = NULL, Volume.saleInfo.listPrice = NULL,
  Volume.saleInfo.offers = NULL, Volume.saleInfo.offers.listPrice = NULL,
  Volume.saleInfo.offers.rentalDuration = NULL,
  Volume.saleInfo.offers.retailPrice = NULL,
  Volume.saleInfo.retailPrice = NULL, Volume.searchInfo = NULL,
  Volume.userInfo = NULL, Volume.userInfo.copy = NULL,
  Volume.userInfo.familySharing = NULL, Volume.userInfo.rentalPeriod = NULL,
  Volume.userInfo.userUploadedVolumeInfo = NULL, Volume.volumeInfo = NULL,
  Volume.volumeInfo.dimensions = NULL, Volume.volumeInfo.imageLinks = NULL,
  Volume.volumeInfo.industryIdentifiers = NULL,
  Volume.volumeInfo.panelizationSummary = NULL, accessInfo = NULL,
  etag = NULL, id = NULL, layerInfo = NULL, recommendedInfo = NULL,
  saleInfo = NULL, searchInfo = NULL, selfLink = NULL, userInfo = NULL,
  volumeInfo = NULL)
}
\arguments{
\item{Volume.accessInfo}{The \link{Volume.accessInfo} object or list of objects}

\item{Volume.accessInfo.epub}{The \link{Volume.accessInfo.epub} object or list of objects}

\item{Volume.accessInfo.pdf}{The \link{Volume.accessInfo.pdf} object or list of objects}

\item{Volume.layerInfo}{The \link{Volume.layerInfo} object or list of objects}

\item{Volume.layerInfo.layers}{The \link{Volume.layerInfo.layers} object or list of objects}

\item{Volume.recommendedInfo}{The \link{Volume.recommendedInfo} object or list of objects}

\item{Volume.saleInfo}{The \link{Volume.saleInfo} object or list of objects}

\item{Volume.saleInfo.listPrice}{The \link{Volume.saleInfo.listPrice} object or list of objects}

\item{Volume.saleInfo.offers}{The \link{Volume.saleInfo.offers} object or list of objects}

\item{Volume.saleInfo.offers.listPrice}{The \link{Volume.saleInfo.offers.listPrice} object or list of objects}

\item{Volume.saleInfo.offers.rentalDuration}{The \link{Volume.saleInfo.offers.rentalDuration} object or list of objects}

\item{Volume.saleInfo.offers.retailPrice}{The \link{Volume.saleInfo.offers.retailPrice} object or list of objects}

\item{Volume.saleInfo.retailPrice}{The \link{Volume.saleInfo.retailPrice} object or list of objects}

\item{Volume.searchInfo}{The \link{Volume.searchInfo} object or list of objects}

\item{Volume.userInfo}{The \link{Volume.userInfo} object or list of objects}

\item{Volume.userInfo.copy}{The \link{Volume.userInfo.copy} object or list of objects}

\item{Volume.userInfo.familySharing}{The \link{Volume.userInfo.familySharing} object or list of objects}

\item{Volume.userInfo.rentalPeriod}{The \link{Volume.userInfo.rentalPeriod} object or list of objects}

\item{Volume.userInfo.userUploadedVolumeInfo}{The \link{Volume.userInfo.userUploadedVolumeInfo} object or list of objects}

\item{Volume.volumeInfo}{The \link{Volume.volumeInfo} object or list of objects}

\item{Volume.volumeInfo.dimensions}{The \link{Volume.volumeInfo.dimensions} object or list of objects}

\item{Volume.volumeInfo.imageLinks}{The \link{Volume.volumeInfo.imageLinks} object or list of objects}

\item{Volume.volumeInfo.industryIdentifiers}{The \link{Volume.volumeInfo.industryIdentifiers} object or list of objects}

\item{Volume.volumeInfo.panelizationSummary}{The \link{Volume.volumeInfo.panelizationSummary} object or list of objects}

\item{accessInfo}{Any information about a volume related to reading or obtaining that volume text}

\item{etag}{Opaque identifier for a specific version of a volume resource}

\item{id}{Unique identifier for a volume}

\item{layerInfo}{What layers exist in this volume and high level information about them}

\item{recommendedInfo}{Recommendation related information for this volume}

\item{saleInfo}{Any information about a volume related to the eBookstore and/or purchaseability}

\item{searchInfo}{Search result information related to this volume}

\item{selfLink}{URL to this resource}

\item{userInfo}{User specific information related to this volume}

\item{volumeInfo}{General volume information}
}
\value{
Volume object
}
\description{
Volume Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Volume functions: \code{\link{Volume.accessInfo.epub}},
  \code{\link{Volume.accessInfo.pdf}},
  \code{\link{Volume.accessInfo}},
  \code{\link{Volume.layerInfo.layers}},
  \code{\link{Volume.layerInfo}},
  \code{\link{Volume.recommendedInfo}},
  \code{\link{Volume.saleInfo.listPrice}},
  \code{\link{Volume.saleInfo.offers.listPrice}},
  \code{\link{Volume.saleInfo.offers.rentalDuration}},
  \code{\link{Volume.saleInfo.offers.retailPrice}},
  \code{\link{Volume.saleInfo.offers}},
  \code{\link{Volume.saleInfo.retailPrice}},
  \code{\link{Volume.saleInfo}},
  \code{\link{Volume.searchInfo}},
  \code{\link{Volume.userInfo.copy}},
  \code{\link{Volume.userInfo.familySharing}},
  \code{\link{Volume.userInfo.rentalPeriod}},
  \code{\link{Volume.userInfo.userUploadedVolumeInfo}},
  \code{\link{Volume.userInfo}},
  \code{\link{Volume.volumeInfo.dimensions}},
  \code{\link{Volume.volumeInfo.imageLinks}},
  \code{\link{Volume.volumeInfo.industryIdentifiers}},
  \code{\link{Volume.volumeInfo.panelizationSummary}},
  \code{\link{Volume.volumeInfo}}
}

