% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Volume.saleInfo.listPrice}
\alias{Volume.saleInfo.listPrice}
\title{Volume.saleInfo.listPrice Object}
\usage{
Volume.saleInfo.listPrice(amount = NULL, currencyCode = NULL)
}
\arguments{
\item{amount}{Amount in the currency listed below}

\item{currencyCode}{An ISO 4217, three-letter currency code}
}
\value{
Volume.saleInfo.listPrice object
}
\description{
Volume.saleInfo.listPrice Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Suggested retail price. (In LITE projection.)
}
\seealso{
Other Volume functions: \code{\link{Volume.accessInfo.epub}},
  \code{\link{Volume.accessInfo.pdf}},
  \code{\link{Volume.accessInfo}},
  \code{\link{Volume.layerInfo.layers}},
  \code{\link{Volume.layerInfo}},
  \code{\link{Volume.recommendedInfo}},
  \code{\link{Volume.saleInfo.offers.listPrice}},
  \code{\link{Volume.saleInfo.offers.rentalDuration}},
  \code{\link{Volume.saleInfo.offers.retailPrice}},
  \code{\link{Volume.saleInfo.offers}},
  \code{\link{Volume.saleInfo.retailPrice}},
  \code{\link{Volume.saleInfo}},
  \code{\link{Volume.searchInfo}},
  \code{\link{Volume.userInfo.copy}},
  \code{\link{Volume.userInfo.familySharing}},
  \code{\link{Volume.userInfo.rentalPeriod}},
  \code{\link{Volume.userInfo.userUploadedVolumeInfo}},
  \code{\link{Volume.userInfo}},
  \code{\link{Volume.volumeInfo.dimensions}},
  \code{\link{Volume.volumeInfo.imageLinks}},
  \code{\link{Volume.volumeInfo.industryIdentifiers}},
  \code{\link{Volume.volumeInfo.panelizationSummary}},
  \code{\link{Volume.volumeInfo}}, \code{\link{Volume}}
}

