% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Volume.saleInfo.offers}
\alias{Volume.saleInfo.offers}
\title{Volume.saleInfo.offers Object}
\usage{
Volume.saleInfo.offers(Volume.saleInfo.offers.listPrice = NULL,
  Volume.saleInfo.offers.rentalDuration = NULL,
  Volume.saleInfo.offers.retailPrice = NULL)
}
\arguments{
\item{Volume.saleInfo.offers.listPrice}{The \link{Volume.saleInfo.offers.listPrice} object or list of objects}

\item{Volume.saleInfo.offers.rentalDuration}{The \link{Volume.saleInfo.offers.rentalDuration} object or list of objects}

\item{Volume.saleInfo.offers.retailPrice}{The \link{Volume.saleInfo.offers.retailPrice} object or list of objects}
}
\value{
Volume.saleInfo.offers object
}
\description{
Volume.saleInfo.offers Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Volume functions: \code{\link{Volume.accessInfo.epub}},
  \code{\link{Volume.accessInfo.pdf}},
  \code{\link{Volume.accessInfo}},
  \code{\link{Volume.layerInfo.layers}},
  \code{\link{Volume.layerInfo}},
  \code{\link{Volume.recommendedInfo}},
  \code{\link{Volume.saleInfo.listPrice}},
  \code{\link{Volume.saleInfo.offers.listPrice}},
  \code{\link{Volume.saleInfo.offers.rentalDuration}},
  \code{\link{Volume.saleInfo.offers.retailPrice}},
  \code{\link{Volume.saleInfo.retailPrice}},
  \code{\link{Volume.saleInfo}},
  \code{\link{Volume.searchInfo}},
  \code{\link{Volume.userInfo.copy}},
  \code{\link{Volume.userInfo.familySharing}},
  \code{\link{Volume.userInfo.rentalPeriod}},
  \code{\link{Volume.userInfo.userUploadedVolumeInfo}},
  \code{\link{Volume.userInfo}},
  \code{\link{Volume.volumeInfo.dimensions}},
  \code{\link{Volume.volumeInfo.imageLinks}},
  \code{\link{Volume.volumeInfo.industryIdentifiers}},
  \code{\link{Volume.volumeInfo.panelizationSummary}},
  \code{\link{Volume.volumeInfo}}, \code{\link{Volume}}
}

