% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Volume.saleInfo}
\alias{Volume.saleInfo}
\title{Volume.saleInfo Object}
\usage{
Volume.saleInfo(Volume.saleInfo.listPrice = NULL,
  Volume.saleInfo.offers = NULL, Volume.saleInfo.offers.listPrice = NULL,
  Volume.saleInfo.offers.rentalDuration = NULL,
  Volume.saleInfo.offers.retailPrice = NULL,
  Volume.saleInfo.retailPrice = NULL, buyLink = NULL, country = NULL,
  isEbook = NULL, listPrice = NULL, offers = NULL, onSaleDate = NULL,
  retailPrice = NULL, saleability = NULL)
}
\arguments{
\item{Volume.saleInfo.listPrice}{The \link{Volume.saleInfo.listPrice} object or list of objects}

\item{Volume.saleInfo.offers}{The \link{Volume.saleInfo.offers} object or list of objects}

\item{Volume.saleInfo.offers.listPrice}{The \link{Volume.saleInfo.offers.listPrice} object or list of objects}

\item{Volume.saleInfo.offers.rentalDuration}{The \link{Volume.saleInfo.offers.rentalDuration} object or list of objects}

\item{Volume.saleInfo.offers.retailPrice}{The \link{Volume.saleInfo.offers.retailPrice} object or list of objects}

\item{Volume.saleInfo.retailPrice}{The \link{Volume.saleInfo.retailPrice} object or list of objects}

\item{buyLink}{URL to purchase this volume on the Google Books site}

\item{country}{The two-letter ISO_3166-1 country code for which this sale information is valid}

\item{isEbook}{Whether or not this volume is an eBook (can be added to the My eBooks shelf)}

\item{listPrice}{Suggested retail price}

\item{offers}{Offers available for this volume (sales and rentals)}

\item{onSaleDate}{The date on which this book is available for sale}

\item{retailPrice}{The actual selling price of the book}

\item{saleability}{Whether or not this book is available for sale or offered for free in the Google eBookstore for the country listed above}
}
\value{
Volume.saleInfo object
}
\description{
Volume.saleInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Any information about a volume related to the eBookstore and/or purchaseability. This information can depend on the country where the request originates from (i.e. books may not be for sale in certain countries).
}
\seealso{
Other Volume functions: \code{\link{Volume.accessInfo.epub}},
  \code{\link{Volume.accessInfo.pdf}},
  \code{\link{Volume.accessInfo}},
  \code{\link{Volume.layerInfo.layers}},
  \code{\link{Volume.layerInfo}},
  \code{\link{Volume.recommendedInfo}},
  \code{\link{Volume.saleInfo.listPrice}},
  \code{\link{Volume.saleInfo.offers.listPrice}},
  \code{\link{Volume.saleInfo.offers.rentalDuration}},
  \code{\link{Volume.saleInfo.offers.retailPrice}},
  \code{\link{Volume.saleInfo.offers}},
  \code{\link{Volume.saleInfo.retailPrice}},
  \code{\link{Volume.searchInfo}},
  \code{\link{Volume.userInfo.copy}},
  \code{\link{Volume.userInfo.familySharing}},
  \code{\link{Volume.userInfo.rentalPeriod}},
  \code{\link{Volume.userInfo.userUploadedVolumeInfo}},
  \code{\link{Volume.userInfo}},
  \code{\link{Volume.volumeInfo.dimensions}},
  \code{\link{Volume.volumeInfo.imageLinks}},
  \code{\link{Volume.volumeInfo.industryIdentifiers}},
  \code{\link{Volume.volumeInfo.panelizationSummary}},
  \code{\link{Volume.volumeInfo}}, \code{\link{Volume}}
}
