% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Volume.userInfo.copy}
\alias{Volume.userInfo.copy}
\title{Volume.userInfo.copy Object}
\usage{
Volume.userInfo.copy(allowedCharacterCount = NULL, limitType = NULL,
  remainingCharacterCount = NULL, updated = NULL)
}
\arguments{
\item{allowedCharacterCount}{No description}

\item{limitType}{No description}

\item{remainingCharacterCount}{No description}

\item{updated}{No description}
}
\value{
Volume.userInfo.copy object
}
\description{
Volume.userInfo.copy Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Copy/Paste accounting information.
}
\seealso{
Other Volume functions: \code{\link{Volume.accessInfo.epub}},
  \code{\link{Volume.accessInfo.pdf}},
  \code{\link{Volume.accessInfo}},
  \code{\link{Volume.layerInfo.layers}},
  \code{\link{Volume.layerInfo}},
  \code{\link{Volume.recommendedInfo}},
  \code{\link{Volume.saleInfo.listPrice}},
  \code{\link{Volume.saleInfo.offers.listPrice}},
  \code{\link{Volume.saleInfo.offers.rentalDuration}},
  \code{\link{Volume.saleInfo.offers.retailPrice}},
  \code{\link{Volume.saleInfo.offers}},
  \code{\link{Volume.saleInfo.retailPrice}},
  \code{\link{Volume.saleInfo}},
  \code{\link{Volume.searchInfo}},
  \code{\link{Volume.userInfo.familySharing}},
  \code{\link{Volume.userInfo.rentalPeriod}},
  \code{\link{Volume.userInfo.userUploadedVolumeInfo}},
  \code{\link{Volume.userInfo}},
  \code{\link{Volume.volumeInfo.dimensions}},
  \code{\link{Volume.volumeInfo.imageLinks}},
  \code{\link{Volume.volumeInfo.industryIdentifiers}},
  \code{\link{Volume.volumeInfo.panelizationSummary}},
  \code{\link{Volume.volumeInfo}}, \code{\link{Volume}}
}

