% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Volume.userInfo.familySharing}
\alias{Volume.userInfo.familySharing}
\title{Volume.userInfo.familySharing Object}
\usage{
Volume.userInfo.familySharing(familyRole = NULL, isSharingAllowed = NULL,
  isSharingDisabledByFop = NULL)
}
\arguments{
\item{familyRole}{The role of the user in the family}

\item{isSharingAllowed}{Whether or not this volume can be shared with the family by the user}

\item{isSharingDisabledByFop}{Whether or not sharing this volume is temporarily disabled due to issues with the Family Wallet}
}
\value{
Volume.userInfo.familySharing object
}
\description{
Volume.userInfo.familySharing Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information on the ability to share with the family.
}
\seealso{
Other Volume functions: \code{\link{Volume.accessInfo.epub}},
  \code{\link{Volume.accessInfo.pdf}},
  \code{\link{Volume.accessInfo}},
  \code{\link{Volume.layerInfo.layers}},
  \code{\link{Volume.layerInfo}},
  \code{\link{Volume.recommendedInfo}},
  \code{\link{Volume.saleInfo.listPrice}},
  \code{\link{Volume.saleInfo.offers.listPrice}},
  \code{\link{Volume.saleInfo.offers.rentalDuration}},
  \code{\link{Volume.saleInfo.offers.retailPrice}},
  \code{\link{Volume.saleInfo.offers}},
  \code{\link{Volume.saleInfo.retailPrice}},
  \code{\link{Volume.saleInfo}},
  \code{\link{Volume.searchInfo}},
  \code{\link{Volume.userInfo.copy}},
  \code{\link{Volume.userInfo.rentalPeriod}},
  \code{\link{Volume.userInfo.userUploadedVolumeInfo}},
  \code{\link{Volume.userInfo}},
  \code{\link{Volume.volumeInfo.dimensions}},
  \code{\link{Volume.volumeInfo.imageLinks}},
  \code{\link{Volume.volumeInfo.industryIdentifiers}},
  \code{\link{Volume.volumeInfo.panelizationSummary}},
  \code{\link{Volume.volumeInfo}}, \code{\link{Volume}}
}

