% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Volume.userInfo}
\alias{Volume.userInfo}
\title{Volume.userInfo Object}
\usage{
Volume.userInfo(Volume.userInfo.copy = NULL,
  Volume.userInfo.familySharing = NULL, Volume.userInfo.rentalPeriod = NULL,
  Volume.userInfo.userUploadedVolumeInfo = NULL, acquiredTime = NULL,
  acquisitionType = NULL, copy = NULL, entitlementType = NULL,
  familySharing = NULL, isFamilySharedFromUser = NULL,
  isFamilySharedToUser = NULL, isFamilySharingAllowed = NULL,
  isFamilySharingDisabledByFop = NULL, isInMyBooks = NULL,
  isPreordered = NULL, isPurchased = NULL, isUploaded = NULL,
  readingPosition = NULL, rentalPeriod = NULL, rentalState = NULL,
  review = NULL, updated = NULL, userUploadedVolumeInfo = NULL)
}
\arguments{
\item{Volume.userInfo.copy}{The \link{Volume.userInfo.copy} object or list of objects}

\item{Volume.userInfo.familySharing}{The \link{Volume.userInfo.familySharing} object or list of objects}

\item{Volume.userInfo.rentalPeriod}{The \link{Volume.userInfo.rentalPeriod} object or list of objects}

\item{Volume.userInfo.userUploadedVolumeInfo}{The \link{Volume.userInfo.userUploadedVolumeInfo} object or list of objects}

\item{acquiredTime}{Timestamp when this volume was acquired by the user}

\item{acquisitionType}{How this volume was acquired}

\item{copy}{Copy/Paste accounting information}

\item{entitlementType}{Whether this volume is purchased, sample, pd download etc}

\item{familySharing}{Information on the ability to share with the family}

\item{isFamilySharedFromUser}{Whether or not the user shared this volume with the family}

\item{isFamilySharedToUser}{Whether or not the user received this volume through family sharing}

\item{isFamilySharingAllowed}{Deprecated: Replaced by familySharing}

\item{isFamilySharingDisabledByFop}{Deprecated: Replaced by familySharing}

\item{isInMyBooks}{Whether or not this volume is currently in 'my books}

\item{isPreordered}{Whether or not this volume was pre-ordered by the authenticated user making the request}

\item{isPurchased}{Whether or not this volume was purchased by the authenticated user making the request}

\item{isUploaded}{Whether or not this volume was user uploaded}

\item{readingPosition}{The user's current reading position in the volume, if one is available}

\item{rentalPeriod}{Period during this book is/was a valid rental}

\item{rentalState}{Whether this book is an active or an expired rental}

\item{review}{This user's review of this volume, if one exists}

\item{updated}{Timestamp when this volume was last modified by a user action, such as a reading position update, volume purchase or writing a review}

\item{userUploadedVolumeInfo}{No description}
}
\value{
Volume.userInfo object
}
\description{
Volume.userInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
User specific information related to this volume. (e.g. page this user last read or whether they purchased this book)
}
\seealso{
Other Volume functions: \code{\link{Volume.accessInfo.epub}},
  \code{\link{Volume.accessInfo.pdf}},
  \code{\link{Volume.accessInfo}},
  \code{\link{Volume.layerInfo.layers}},
  \code{\link{Volume.layerInfo}},
  \code{\link{Volume.recommendedInfo}},
  \code{\link{Volume.saleInfo.listPrice}},
  \code{\link{Volume.saleInfo.offers.listPrice}},
  \code{\link{Volume.saleInfo.offers.rentalDuration}},
  \code{\link{Volume.saleInfo.offers.retailPrice}},
  \code{\link{Volume.saleInfo.offers}},
  \code{\link{Volume.saleInfo.retailPrice}},
  \code{\link{Volume.saleInfo}},
  \code{\link{Volume.searchInfo}},
  \code{\link{Volume.userInfo.copy}},
  \code{\link{Volume.userInfo.familySharing}},
  \code{\link{Volume.userInfo.rentalPeriod}},
  \code{\link{Volume.userInfo.userUploadedVolumeInfo}},
  \code{\link{Volume.volumeInfo.dimensions}},
  \code{\link{Volume.volumeInfo.imageLinks}},
  \code{\link{Volume.volumeInfo.industryIdentifiers}},
  \code{\link{Volume.volumeInfo.panelizationSummary}},
  \code{\link{Volume.volumeInfo}}, \code{\link{Volume}}
}

