% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Volume.volumeInfo.dimensions}
\alias{Volume.volumeInfo.dimensions}
\title{Volume.volumeInfo.dimensions Object}
\usage{
Volume.volumeInfo.dimensions(height = NULL, thickness = NULL,
  width = NULL)
}
\arguments{
\item{height}{Height or length of this volume (in cm)}

\item{thickness}{Thickness of this volume (in cm)}

\item{width}{Width of this volume (in cm)}
}
\value{
Volume.volumeInfo.dimensions object
}
\description{
Volume.volumeInfo.dimensions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Physical dimensions of this volume.
}
\seealso{
Other Volume functions: \code{\link{Volume.accessInfo.epub}},
  \code{\link{Volume.accessInfo.pdf}},
  \code{\link{Volume.accessInfo}},
  \code{\link{Volume.layerInfo.layers}},
  \code{\link{Volume.layerInfo}},
  \code{\link{Volume.recommendedInfo}},
  \code{\link{Volume.saleInfo.listPrice}},
  \code{\link{Volume.saleInfo.offers.listPrice}},
  \code{\link{Volume.saleInfo.offers.rentalDuration}},
  \code{\link{Volume.saleInfo.offers.retailPrice}},
  \code{\link{Volume.saleInfo.offers}},
  \code{\link{Volume.saleInfo.retailPrice}},
  \code{\link{Volume.saleInfo}},
  \code{\link{Volume.searchInfo}},
  \code{\link{Volume.userInfo.copy}},
  \code{\link{Volume.userInfo.familySharing}},
  \code{\link{Volume.userInfo.rentalPeriod}},
  \code{\link{Volume.userInfo.userUploadedVolumeInfo}},
  \code{\link{Volume.userInfo}},
  \code{\link{Volume.volumeInfo.imageLinks}},
  \code{\link{Volume.volumeInfo.industryIdentifiers}},
  \code{\link{Volume.volumeInfo.panelizationSummary}},
  \code{\link{Volume.volumeInfo}}, \code{\link{Volume}}
}
