% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Volume.volumeInfo.imageLinks}
\alias{Volume.volumeInfo.imageLinks}
\title{Volume.volumeInfo.imageLinks Object}
\usage{
Volume.volumeInfo.imageLinks(extraLarge = NULL, large = NULL,
  medium = NULL, small = NULL, smallThumbnail = NULL, thumbnail = NULL)
}
\arguments{
\item{extraLarge}{Image link for extra large size (width of ~1280 pixels)}

\item{large}{Image link for large size (width of ~800 pixels)}

\item{medium}{Image link for medium size (width of ~575 pixels)}

\item{small}{Image link for small size (width of ~300 pixels)}

\item{smallThumbnail}{Image link for small thumbnail size (width of ~80 pixels)}

\item{thumbnail}{Image link for thumbnail size (width of ~128 pixels)}
}
\value{
Volume.volumeInfo.imageLinks object
}
\description{
Volume.volumeInfo.imageLinks Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of image links for all the sizes that are available. (In LITE projection.)
}
\seealso{
Other Volume functions: \code{\link{Volume.accessInfo.epub}},
  \code{\link{Volume.accessInfo.pdf}},
  \code{\link{Volume.accessInfo}},
  \code{\link{Volume.layerInfo.layers}},
  \code{\link{Volume.layerInfo}},
  \code{\link{Volume.recommendedInfo}},
  \code{\link{Volume.saleInfo.listPrice}},
  \code{\link{Volume.saleInfo.offers.listPrice}},
  \code{\link{Volume.saleInfo.offers.rentalDuration}},
  \code{\link{Volume.saleInfo.offers.retailPrice}},
  \code{\link{Volume.saleInfo.offers}},
  \code{\link{Volume.saleInfo.retailPrice}},
  \code{\link{Volume.saleInfo}},
  \code{\link{Volume.searchInfo}},
  \code{\link{Volume.userInfo.copy}},
  \code{\link{Volume.userInfo.familySharing}},
  \code{\link{Volume.userInfo.rentalPeriod}},
  \code{\link{Volume.userInfo.userUploadedVolumeInfo}},
  \code{\link{Volume.userInfo}},
  \code{\link{Volume.volumeInfo.dimensions}},
  \code{\link{Volume.volumeInfo.industryIdentifiers}},
  \code{\link{Volume.volumeInfo.panelizationSummary}},
  \code{\link{Volume.volumeInfo}}, \code{\link{Volume}}
}

