% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Volume.volumeInfo}
\alias{Volume.volumeInfo}
\title{Volume.volumeInfo Object}
\usage{
Volume.volumeInfo(Volume.volumeInfo.dimensions = NULL,
  Volume.volumeInfo.imageLinks = NULL,
  Volume.volumeInfo.industryIdentifiers = NULL,
  Volume.volumeInfo.panelizationSummary = NULL, allowAnonLogging = NULL,
  authors = NULL, averageRating = NULL, canonicalVolumeLink = NULL,
  categories = NULL, contentVersion = NULL, description = NULL,
  dimensions = NULL, imageLinks = NULL, industryIdentifiers = NULL,
  infoLink = NULL, language = NULL, mainCategory = NULL,
  maturityRating = NULL, pageCount = NULL, panelizationSummary = NULL,
  previewLink = NULL, printType = NULL, printedPageCount = NULL,
  publishedDate = NULL, publisher = NULL, ratingsCount = NULL,
  readingModes = NULL, samplePageCount = NULL, seriesInfo = NULL,
  subtitle = NULL, title = NULL)
}
\arguments{
\item{Volume.volumeInfo.dimensions}{The \link{Volume.volumeInfo.dimensions} object or list of objects}

\item{Volume.volumeInfo.imageLinks}{The \link{Volume.volumeInfo.imageLinks} object or list of objects}

\item{Volume.volumeInfo.industryIdentifiers}{The \link{Volume.volumeInfo.industryIdentifiers} object or list of objects}

\item{Volume.volumeInfo.panelizationSummary}{The \link{Volume.volumeInfo.panelizationSummary} object or list of objects}

\item{allowAnonLogging}{Whether anonymous logging should be allowed}

\item{authors}{The names of the authors and/or editors for this volume}

\item{averageRating}{The mean review rating for this volume}

\item{canonicalVolumeLink}{Canonical URL for a volume}

\item{categories}{A list of subject categories, such as 'Fiction', 'Suspense', etc}

\item{contentVersion}{An identifier for the version of the volume content (text & images)}

\item{description}{A synopsis of the volume}

\item{dimensions}{Physical dimensions of this volume}

\item{imageLinks}{A list of image links for all the sizes that are available}

\item{industryIdentifiers}{Industry standard identifiers for this volume}

\item{infoLink}{URL to view information about this volume on the Google Books site}

\item{language}{Best language for this volume (based on content)}

\item{mainCategory}{The main category to which this volume belongs}

\item{maturityRating}{No description}

\item{pageCount}{Total number of pages as per publisher metadata}

\item{panelizationSummary}{A top-level summary of the panelization info in this volume}

\item{previewLink}{URL to preview this volume on the Google Books site}

\item{printType}{Type of publication of this volume}

\item{printedPageCount}{Total number of printed pages in generated pdf representation}

\item{publishedDate}{Date of publication}

\item{publisher}{Publisher of this volume}

\item{ratingsCount}{The number of review ratings for this volume}

\item{readingModes}{The reading modes available for this volume}

\item{samplePageCount}{Total number of sample pages as per publisher metadata}

\item{seriesInfo}{No description}

\item{subtitle}{Volume subtitle}

\item{title}{Volume title}
}
\value{
Volume.volumeInfo object
}
\description{
Volume.volumeInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
General volume information.
}
\seealso{
Other Volume functions: \code{\link{Volume.accessInfo.epub}},
  \code{\link{Volume.accessInfo.pdf}},
  \code{\link{Volume.accessInfo}},
  \code{\link{Volume.layerInfo.layers}},
  \code{\link{Volume.layerInfo}},
  \code{\link{Volume.recommendedInfo}},
  \code{\link{Volume.saleInfo.listPrice}},
  \code{\link{Volume.saleInfo.offers.listPrice}},
  \code{\link{Volume.saleInfo.offers.rentalDuration}},
  \code{\link{Volume.saleInfo.offers.retailPrice}},
  \code{\link{Volume.saleInfo.offers}},
  \code{\link{Volume.saleInfo.retailPrice}},
  \code{\link{Volume.saleInfo}},
  \code{\link{Volume.searchInfo}},
  \code{\link{Volume.userInfo.copy}},
  \code{\link{Volume.userInfo.familySharing}},
  \code{\link{Volume.userInfo.rentalPeriod}},
  \code{\link{Volume.userInfo.userUploadedVolumeInfo}},
  \code{\link{Volume.userInfo}},
  \code{\link{Volume.volumeInfo.dimensions}},
  \code{\link{Volume.volumeInfo.imageLinks}},
  \code{\link{Volume.volumeInfo.industryIdentifiers}},
  \code{\link{Volume.volumeInfo.panelizationSummary}},
  \code{\link{Volume}}
}

