% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Volume2}
\alias{Volume2}
\title{Volume2 Object}
\usage{
Volume2(items = NULL, nextPageToken = NULL)
}
\arguments{
\item{items}{A list of volumes}

\item{nextPageToken}{No description}
}
\value{
Volume2 object
}
\description{
Volume2 Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

