% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Volumeannotation.contentRanges}
\alias{Volumeannotation.contentRanges}
\title{Volumeannotation.contentRanges Object}
\usage{
Volumeannotation.contentRanges(cfiRange = NULL, contentVersion = NULL,
  gbImageRange = NULL, gbTextRange = NULL)
}
\arguments{
\item{cfiRange}{Range in CFI format for this annotation for version above}

\item{contentVersion}{Content version applicable to ranges below}

\item{gbImageRange}{Range in GB image format for this annotation for version above}

\item{gbTextRange}{Range in GB text format for this annotation for version above}
}
\value{
Volumeannotation.contentRanges object
}
\description{
Volumeannotation.contentRanges Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The content ranges to identify the selected text.
}
\seealso{
Other Volumeannotation functions: \code{\link{Volumeannotation}}
}

