% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Volumeannotation}
\alias{Volumeannotation}
\title{Volumeannotation Object}
\usage{
Volumeannotation(Volumeannotation.contentRanges = NULL,
  annotationDataId = NULL, annotationDataLink = NULL,
  annotationType = NULL, contentRanges = NULL, data = NULL,
  deleted = NULL, id = NULL, layerId = NULL, pageIds = NULL,
  selectedText = NULL, selfLink = NULL, updated = NULL, volumeId = NULL)
}
\arguments{
\item{Volumeannotation.contentRanges}{The \link{Volumeannotation.contentRanges} object or list of objects}

\item{annotationDataId}{The annotation data id for this volume annotation}

\item{annotationDataLink}{Link to get data for this annotation}

\item{annotationType}{The type of annotation this is}

\item{contentRanges}{The content ranges to identify the selected text}

\item{data}{Data for this annotation}

\item{deleted}{Indicates that this annotation is deleted}

\item{id}{Unique id of this volume annotation}

\item{layerId}{The Layer this annotation is for}

\item{pageIds}{Pages the annotation spans}

\item{selectedText}{Excerpt from the volume}

\item{selfLink}{URL to this resource}

\item{updated}{Timestamp for the last time this anntoation was updated}

\item{volumeId}{The Volume this annotation is for}
}
\value{
Volumeannotation object
}
\description{
Volumeannotation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Volumeannotation functions: \code{\link{Volumeannotation.contentRanges}}
}
