% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Volumeannotations}
\alias{Volumeannotations}
\title{Volumeannotations Object}
\usage{
Volumeannotations(items = NULL, nextPageToken = NULL, totalItems = NULL,
  version = NULL)
}
\arguments{
\item{items}{A list of volume annotations}

\item{nextPageToken}{Token to pass in for pagination for the next page}

\item{totalItems}{The total number of volume annotations found}

\item{version}{The version string for all of the volume annotations in this layer (not just the ones in this response)}
}
\value{
Volumeannotations object
}
\description{
Volumeannotations Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
