% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_functions.R
\name{volumes.get}
\alias{volumes.get}
\title{Gets volume information for a single volume.}
\usage{
volumes.get(volumeId, country = NULL, includeNonComicsSeries = NULL,
  partner = NULL, projection = NULL, source = NULL,
  user_library_consistent_read = NULL)
}
\arguments{
\item{volumeId}{ID of volume to retrieve}

\item{country}{ISO-3166-1 code to override the IP-based location}

\item{includeNonComicsSeries}{Set to true to include non-comics series}

\item{partner}{Brand results for partner ID}

\item{projection}{Restrict information returned to a set of selected fields}

\item{source}{String to identify the originator of this request}

\item{user_library_consistent_read}{}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/books
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/books)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/books/docs/v1/getting_started}{Google Documentation}
}

