% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Volumes}
\alias{Volumes}
\title{Volumes Object}
\usage{
Volumes(items = NULL, totalItems = NULL)
}
\arguments{
\item{items}{A list of volumes}

\item{totalItems}{Total number of volumes found}
}
\value{
Volumes object
}
\description{
Volumes Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
