% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Volumeseriesinfo}
\alias{Volumeseriesinfo}
\title{Volumeseriesinfo Object}
\usage{
Volumeseriesinfo(Volumeseriesinfo.volumeSeries = NULL,
  Volumeseriesinfo.volumeSeries.issue = NULL, bookDisplayNumber = NULL,
  shortSeriesBookTitle = NULL, volumeSeries = NULL)
}
\arguments{
\item{Volumeseriesinfo.volumeSeries}{The \link{Volumeseriesinfo.volumeSeries} object or list of objects}

\item{Volumeseriesinfo.volumeSeries.issue}{The \link{Volumeseriesinfo.volumeSeries.issue} object or list of objects}

\item{bookDisplayNumber}{The display number string}

\item{shortSeriesBookTitle}{Short book title in the context of the series}

\item{volumeSeries}{No description}
}
\value{
Volumeseriesinfo object
}
\description{
Volumeseriesinfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Volumeseriesinfo functions: \code{\link{Volumeseriesinfo.volumeSeries.issue}},
  \code{\link{Volumeseriesinfo.volumeSeries}}
}
