% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_objects.R
\name{Volumeseriesinfo.volumeSeries}
\alias{Volumeseriesinfo.volumeSeries}
\title{Volumeseriesinfo.volumeSeries Object}
\usage{
Volumeseriesinfo.volumeSeries(Volumeseriesinfo.volumeSeries.issue = NULL)
}
\arguments{
\item{Volumeseriesinfo.volumeSeries.issue}{The \link{Volumeseriesinfo.volumeSeries.issue} object or list of objects}
}
\value{
Volumeseriesinfo.volumeSeries object
}
\description{
Volumeseriesinfo.volumeSeries Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Volumeseriesinfo functions: \code{\link{Volumeseriesinfo.volumeSeries.issue}},
  \code{\link{Volumeseriesinfo}}
}
